Added softmax + still not training - softmax added as operation in ops - in convolution backward only inputs had backward_fn and if inputs had requires_grad=False, then it wouldn't execute, hence split the gradient calc into inputs_backward, kernel_backward and bias_backward - still not training, will try to debug later or tmrw

+++ neograd/autograd/utils.py
@@ -15,7 +15,7 @@ def process_data(data):
     if not isinstance(data, np.ndarray):
       data = np.array(data)
     try:
-      data = data.astype(float)
+      data = data.astype(np.float32)
     except ValueError:
       raise TypeError("Elements of data should be of type float or be typecastable to float")
   else:
